BackEnd/requirements.txt
Django
django-rest-framework
drf-yasg
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
environs